services/app/requirements.txt
# EnergyPulse App Service Dependencies

# Core
psycopg2-binary==2.9.9
pyyaml==6.0.2
requests==2.32.3
apscheduler==3.10.4

# Data Processing
pandas==2.2.3
numpy==1.26.4

# Chinese A-Share Data
tushare==1.4.19
akshare==1.18.52

# News/RSS
feedparser==6.0.11

# Markdown → HTML
markdown==3.7
jinja2==3.1.4

# Utilities
python-dateutil==2.9.0
pytz==2024.1
yfinance==0.2.37
